requirements.txt
torch
transformers>=4.30.0
datasets>=2.12.0
peft>=0.4.0
accelerate>=0.20.0
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.24.0
flask>=2.3.0
wandb>=0.15.0
praw>=7.7.0
newspaper3k>=0.2.8
scikit-learn>=1.3.0
nltk>=3.8.0
jupyter
ipykernel
pytest
flake8
tqdm
python-dotenv
wikipedia
requests
beautifulsoup4